refactor: Use JSON parser for new SchemaDef
