+++ tools/schema/main.go
@@ -196,7 +196,7 @@ func generateSchemaNew() error {
 	schemaDef.Typedefs = make(model.DefMap)
 	schemaDef.State = make(model.DefMap)
 
-	defMapKey := &model.DefElt{Val: "owner"}
+	defMapKey := model.DefElt{Val: "owner"}
 	schemaDef.State[defMapKey] = &model.DefElt{Val: "AgentID // current owner of this smart contract"}
 	schemaDef.Funcs = make(model.FuncDefMap)
 	schemaDef.Views = make(model.FuncDefMap)
@@ -230,8 +230,9 @@ func loadSchema(file *os.File) (s *model.Schema, err error) {
 	schemaDef := &model.SchemaDef{}
 	switch filepath.Ext(file.Name()) {
 	case ".json":
-		// FIXME support JSON at the same time
-		err = json.NewDecoder(file).Decode(schemaDef)
+		var jsonSchemaDef model.JSONSchemaDef
+		err = json.NewDecoder(file).Decode(&jsonSchemaDef)
+		schemaDef = jsonSchemaDef.ToSchemaDef()
 		if err == nil && *flagType == "convert" {
 			err = WriteYAMLSchema(schemaDef)
 		}
@@ -265,6 +266,7 @@ func WriteJSONSchema(schemaDef *model.SchemaDef) error {
 	}
 	defer file.Close()
 
+	// FIXME convert to `JSONSchemaDef` first
 	b, err := json.Marshal(schemaDef)
 	if err != nil {
 		return err
@@ -287,6 +289,7 @@ func WriteYAMLSchema(schemaDef *model.SchemaDef) error {
 	}
 	defer file.Close()
 
+	// FIXME convert to `YAMLSchemaDef` first
 	b, err := yaml.Marshal(schemaDef)
 	if err != nil {
 		return err
