--- tools/schema/model/json_schema.go
@@ -1,1 +1,0 @@
-package model
