+++ tools/schema/model/schema.go
@@ -7,94 +7,9 @@ import (
 	"fmt"
 	"sort"
 	"strings"
-	"time"
 
 	"github.com/iotaledger/wasp/packages/iscp"
 )
-
-// TODO describe schema details in docs
-type (
-	FieldMap      map[string]*Field
-	FieldMapMap   map[string]FieldMap
-	StringMap     map[string]string
-	StringMapMap  map[string]StringMap
-	DefMap        map[*DefElt]*DefElt
-	DefNameMap    map[string]*DefElt
-	DefMapMap     map[*DefElt]*DefMap
-	DefNameMapMap map[string]*DefMap
-)
-
-type FuncDef struct {
-	Access  DefElt
-	Params  DefMap
-	Results DefMap
-	Line    int
-	Comment string
-}
-type FuncDefMap map[string]*FuncDef
-
-type DefElt struct {
-	Val     string
-	Comment string
-	Line    int
-}
-
-type SchemaDef struct {
-	Name        DefElt
-	Description DefElt
-	Events      DefMapMap
-	Structs     DefMapMap
-	Typedefs    DefMap
-	State       DefMap
-	Funcs       FuncDefMap
-	Views       FuncDefMap
-}
-
-func NewSchemaDef() *SchemaDef {
-	def := &SchemaDef{}
-	def.Events = make(DefMapMap)
-	def.Structs = make(DefMapMap)
-	def.Typedefs = make(DefMap)
-	def.State = make(DefMap)
-	def.Funcs = make(FuncDefMap)
-	def.Views = make(FuncDefMap)
-	return def
-}
-
-type Func struct {
-	Name    string
-	Access  DefElt
-	Kind    string
-	Hname   iscp.Hname
-	Params  []*Field
-	Results []*Field
-	Line    int
-	Comment string
-}
-
-type Struct struct {
-	Name   DefElt
-	Fields []*Field
-}
-
-type Schema struct {
-	ContractName  string
-	PackageName   string
-	Description   string
-	CoreContracts bool
-	SchemaTime    time.Time
-	Events        []*Struct
-	Funcs         []*Func
-	Params        []*Field
-	Results       []*Field
-	StateVars     []*Field
-	Structs       []*Struct
-	Typedefs      []*Field
-}
-
-func NewSchema() *Schema {
-	return &Schema{}
-}
 
 func (s *Schema) Compile(schemaDef *SchemaDef) error {
 	s.ContractName = strings.TrimSpace(schemaDef.Name.Val)
@@ -137,7 +52,7 @@ func (s *Schema) Compile(schemaDef *SchemaDef) error {
 
 func (s *Schema) compileEvents(schemaDef *SchemaDef) error {
 	for _, eventName := range sortedMaps(schemaDef.Events) {
-		event, err := s.compileStruct("event", *eventName, *schemaDef.Events[eventName])
+		event, err := s.compileStruct("event", eventName, *schemaDef.Events[eventName])
 		if err != nil {
 			return err
 		}
@@ -206,7 +121,7 @@ func (s *Schema) compileFuncFields(fieldMap DefMap, allFieldMap *FieldMap, what 
 	fieldAliases := make(DefNameMap)
 	for _, fldName := range sortedKeys(fieldMap) {
 		fldType := fieldMap[fldName]
-		field, err := s.compileField(fldName, fldType)
+		field, err := s.compileField(&fldName, fldType)
 		if err != nil {
 			return nil, err
 		}
@@ -249,7 +164,7 @@ func (s *Schema) compileStateVars(schemaDef *SchemaDef) error {
 	varAliases := make(DefNameMap)
 	for _, varName := range sortedKeys(schemaDef.State) {
 		varType := schemaDef.State[varName]
-		varDef, err := s.compileField(varName, varType)
+		varDef, err := s.compileField(&varName, varType)
 		if err != nil {
 			return err
 		}
@@ -278,7 +193,7 @@ func (s *Schema) compileStateVars(schemaDef *SchemaDef) error {
 
 func (s *Schema) compileStructs(schemaDef *SchemaDef) error {
 	for _, structName := range sortedMaps(schemaDef.Structs) {
-		structDef, err := s.compileStruct("struct", *structName, *schemaDef.Structs[structName])
+		structDef, err := s.compileStruct("struct", structName, *schemaDef.Structs[structName])
 		if err != nil {
 			return err
 		}
@@ -293,7 +208,7 @@ func (s *Schema) compileStruct(kind string, structName DefElt, structFields DefM
 	fieldAliases := make(DefNameMap)
 	for _, fldName := range sortedKeys(structFields) {
 		fldType := structFields[fldName]
-		field, err := s.compileField(fldName, fldType)
+		field, err := s.compileField(&fldName, fldType)
 		if err != nil {
 			return nil, err
 		}
@@ -332,7 +247,7 @@ func (s *Schema) compileTypeDefs(schemaDef *SchemaDef) error {
 	varAliases := make(DefNameMap)
 	for _, varName := range sortedKeys(schemaDef.Typedefs) {
 		varType := schemaDef.Typedefs[varName]
-		varDef, err := s.compileField(varName, varType)
+		varDef, err := s.compileField(&varName, varType)
 		if err != nil {
 			return err
 		}
@@ -375,13 +290,13 @@ func sortedFuncDescs(dict FuncDefMap) []string {
 	return keys
 }
 
-type DefEltList []*DefElt
+type DefEltList []DefElt
 
 func (l DefEltList) Len() int           { return len(l) }
 func (l DefEltList) Swap(i, j int)      { l[i], l[j] = l[j], l[i] }
 func (l DefEltList) Less(i, j int) bool { return l[i].Val < l[j].Val }
 
-func sortedKeys(dict DefMap) []*DefElt {
+func sortedKeys(dict DefMap) []DefElt {
 	keys := make(DefEltList, 0)
 	for key, _ := range dict {
 		keys = append(keys, key)
@@ -390,7 +305,7 @@ func sortedKeys(dict DefMap) []*DefElt {
 	return keys
 }
 
-func sortedMaps(dict DefMapMap) []*DefElt {
+func sortedMaps(dict DefMapMap) []DefElt {
 	keys := make(DefEltList, 0)
 	for key := range dict {
 		keys = append(keys, key)
